// src/api/generated/models/billableLine.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { BillableLineQuoteLineSummary } from './billableLineQuoteLineSummary';
import type { BillableLineRequestLine } from './billableLineRequestLine';

export interface BillableLine {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  selectionLineId: string;
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  lineageId: string;
  quantity: string;
  selectionQuantity: string;
  invoicedQuantity: string;
  remainingQuantity: string;
  unitPrice: string;
  lineTotal: string;
  quoteLine: BillableLineQuoteLineSummary;
  requestLine: BillableLineRequestLine;
}